scripts/vector-search/requirements.txt
openai>=1.3.0
supabase>=2.0.0
tiktoken>=0.5.0
python-docx>=0.8.11
pypdf>=3.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
